Moved unused files to a legacy code folder. Some of the work in there may be reusable later but for now it is not being used or maintained. Cleaned up several files and deleted some files that will no longer be used at all
