+++ legacyCode/legacySERVER/models/Class.js
