+++ legacyCode/legacySERVER/models/Spell.js
