+++ legacyCode/legacySERVER/routes/spellRoutes.js
