+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import Nav from "./components/Nav/index";
 import "./App.css";
 import Landing from "./pages/Landing.js";
@@ -8,7 +8,6 @@ import {
   Route,
   Redirect,
 } from "react-router-dom";
-import Spellbook from "./pages/SpellBook";
 import Characters from "./pages/Characters"
 import PlayerDisplay from "./pages/PlayerDisplay";
 import ComponentTest from "./pages/ComponentTest";
@@ -35,9 +34,6 @@ function App() {
             <PlayerDisplay />) : (
             <Redirect to="/login" />
           )}
-        </Route>
-        <Route exact path="/spellbook">
-          <Spellbook />
         </Route>
         <Route exact path="/componenttest">
           <ComponentTest />
